# src/main/docs/api.yaml
openapi: 3.0.0
info:
  description: SelfEducation
  version: 1.0.5
  title: SelfEducation
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/desoluy/SelfEducation/1.0.5
  - url: 'https://petstore.swagger.io/v2'
paths:
...
  /theme/{id}/images:
    get:
      summary: Retrieve images for theme
      operationId: themeImages
      parameters:
        - name: token
          in: header
          description: Authentication token
          required: true
          style: simple
          explode: true
          schema:
            type: string
        - name: id
          in: path
          description: Theme id
          required: true
          style: simple
          explode: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successfully retrieved images for theme
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImageMaterial'
        "400":
          description: Invalid theme id was provided
        "404":
          description: Images weren't found for theme
  /theme/{id}/videos:
    get:
      summary: Retrieve videos for theme
      operationId: themeVideos
      parameters:
        - name: token
          in: header
          description: Authentication token
          required: true
          style: simple
          explode: true
          schema:
            type: string
        - name: id
          in: path
          description: Theme id
          required: true
          style: simple
          explode: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successfully retrieved videos for theme
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VideoMaterial'
        "400":
          description: Invalid theme id was provided
        "404":
          description: Videos weren't found for theme
  /theme/{id}/tasks:
    get:
      summary: Retrieve tasks for theme
      operationId: themeTasks
      parameters:
        - name: token
          in: header
          description: Authentication token
          required: true
          style: simple
          explode: true
          schema:
            type: string
        - name: id
          in: path
          description: Theme id
          required: true
          style: simple
          explode: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successfully retrieved tasks for theme
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreativeTask'
        "400":
          description: Invalid theme id was provided
        "404":
          description: Tasks weren't found for theme
  /theme/{id}/tests:
    get:
      summary: Retrieve tests for theme
      operationId: themeTests
      parameters:
        - name: token
          in: header
          description: Authentication token
          required: true
          style: simple
          explode: true
          schema:
            type: string
        - name: id
          in: path
          description: Theme id
          required: true
          style: simple
          explode: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successfully retrieved tests for theme
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Test'
        "400":
          description: Invalid theme id was provided
        "404":
          description: Tests weren't found for theme
  /theme/{id}/texts:
    get:
      summary: Retrieve texts for theme
      operationId: themeTexts
      parameters:
        - name: token
          in: header
          description: Authentication token
          required: true
          style: simple
          explode: true
          schema:
            type: string
        - name: id
          in: path
          description: Theme id
          required: true
          style: simple
          explode: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successfully retrieved texts for theme
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/TextMaterial'
        "400":
          description: Invalid theme id was provided
        "404":
          description: Texts weren't found for theme
  /theme/{id}/statuses:
    get:
      summary: Retrieve statuses for theme (admin only)
      operationId: themeStatuses
      parameters:
        - name: token
          in: header
          description: Authentication token
          required: true
          style: simple
          explode: true
          schema:
            type: string
        - name: id
          in: path
          description: Theme id
          required: true
          style: simple
          explode: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successfully retrieved statuses for theme
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ThemeStatus'
        "400":
          description: Invalid theme id was provided
        "404":
          description: Statuses weren't found for theme
...
components:
  schemas:
...
    ImageMaterial:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: false
        name:
          type: string
        priority:
          type: integer
          format: int64
          nullable: false
        url:
          type: string
          nullable: false
      required:
        - id
        - name
        - priority
        - url
    VideoMaterial:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: false
        name:
          type: string
        priority:
          type: integer
          format: int64
          nullable: false
        url:
          type: string
          nullable: false
      required:
        - id
        - name
        - priority
        - url
    CreativeTask:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: false
        name:
          type: string
        priority:
          type: integer
          format: int64
          nullable: false
        text:
          type: string
          nullable: false
      required:
        - id
        - name
        - priority
        - text
...
    Test:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: false
        name:
          type: string
          nullable: false
        priority:
          type: integer
          format: int64
          nullable: false
      required:
        - id
        - name
        - priority
...
    TextMaterial:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: false
        name:
          type: string
        priority:
          type: integer
          format: int64
          nullable: false
        text:
          type: string
          nullable: false
      required:
        - id
        - name
        - priority
        - text
    ThemeStatus:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: false
        wasSeen:
          type: boolean
          nullable: false
        participant:
          type: integer
          format: int64
          description: User id
          nullable: false
      required:
        - id
        - wasSeen
        - participant
...